+++ neoforge/src/main/java/com/clefal/lootbeams/LootBeamsForgeModClientEvent.java
@@ -4,6 +4,7 @@ package com.clefal.lootbeams;
 import com.clefal.lootbeams.compat.AccessoriesCompatModule;
 import com.clefal.lootbeams.compat.ApotheosisCompatModule;
 import com.clefal.lootbeams.compat.CuriosContinuationAndAdornedCompatModule;
+import com.clefal.lootbeams.compat.TieredReforgedCompatModule;
 import com.clefal.lootbeams.config.ConfigHandlers;
 import com.clefal.lootbeams.modules.ModulesManager;
 import com.clefal.lootbeams.modules.tooltip.overlay.AdvanceTooltipOverlay;
@@ -33,7 +34,8 @@ public class LootBeamsForgeModClientEvent {
         ModulesManager.registerModules(
                 CuriosContinuationAndAdornedCompatModule.INSTANCE,
                 AccessoriesCompatModule.INSTANCE,
-                ApotheosisCompatModule.INSTANCE
+                ApotheosisCompatModule.INSTANCE,
+                TieredReforgedCompatModule.INSTANCE
         );
         ModulesManager.enableAll();
         ConfigHandlers.init();